fix destory entry bug

+++ app/controllers/kpi_entries_controller.rb
@@ -53,16 +53,18 @@ class KpiEntriesController < ApplicationController
     msg = Message.new
     msg.result = true
     entry = KpiEntry.find_by(id:params[:id])
-    if entry.entry_type == 1
-      msg.result =false
-      msg.content = I18n.t "entry.desc.del-error"
-    else
-      entry.destroy
-      total = KpiEntry.find_by(parsed_entry_at:entry.parsed_entry_at,entity_id:entry.entity_id,user_kpi_item_id:entry.user_kpi_item_id,entry_type:1)
-      if total
-        msg.content = {item_id:entry.user_kpi_item_id,value:total.value}
+    if entry
+      if entry.entry_type == 1
+        msg.result =false
+        msg.content = I18n.t "entry.desc.del-error"
       else
-        msg.content = {item_id:entry.user_kpi_item_id,value:""}
+        entry.destroy
+        total = KpiEntry.find_by(parsed_entry_at:entry.parsed_entry_at,entity_id:entry.entity_id,user_kpi_item_id:entry.user_kpi_item_id,entry_type:1)
+        if total
+          msg.content = {item_id:entry.user_kpi_item_id,value:total.value}
+        else
+          msg.content = {item_id:entry.user_kpi_item_id,value:""}
+        end
       end
     end
     render :json=> msg
